Service adicionado

+++ src/main/java/br/uniesp/si/techback/repository/FilmeRepository.java
@@ -3,5 +3,5 @@ package br.uniesp.si.techback.repository;
 import br.uniesp.si.techback.model.Filme;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FilmeRepository extends JpaRepository<Filme, Integer> {
+public interface FilmeRepository extends JpaRepository<Filme, Long> {
 }
